move data_structure_dicts, obsolete_variables and variable_metadata to process/core/data_structure
