+++ process/core/data_structure/obsolete_vars.py
