+++ process/core/io/in_dat/base.py
@@ -13,8 +13,8 @@ Notes
 import sys
 from re import sub
 
+from process.core.data_structure.data_structure_dicts import get_dicts
 from process.core.exceptions import ProcessValidationError
-from process.core.io.data_structure_dicts import get_dicts
 from process.core.solver.constraints import ConstraintManager
 from process.core.solver.iteration_variables import ITERATION_VARIABLES
 from process.data_structure.numerics import PROCESSRunMode
