+++ process/core/io/plot/scans.py
@@ -35,8 +35,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.ticker import MultipleLocator, PercentFormatter
 
+from process.core.data_structure.variable_metadata import var_dicts as meta
 from process.core.io.mfile import MFile
-from process.core.io.variable_metadata import var_dicts as meta
 from process.core.scan import ScanVariables
 from process.data_structure.numerics import SolverOutputCondition
